chore(api): split schema and ai model primitives
