+++ src/uw_scan/models/trade_insights_ai_parts/__init__.py
@@ -1,0 +1,1 @@
+"""Internal building blocks for Trade Insights AI contracts."""
